Create basic_LTM_folder.zip
Added a new example.

This example reads input data from a temporary working directory created by a previous pyProdrisk run (in this example: a run with the basic.py example).

This might be adjusted to read data from an already existing v10 EOPS dataset, which might be the current data storage for ProdRisk data for the model users?