Changed temp-dir path of all examples
Commented out prodrisk.keep_working_directory, as this should be considered a debug-option. Might also be a temporary solution for users depending on SINTEF result applications such as pckurvetegn and ET.

Added prodrisk.temp_dir = "C:\\temp\\" to all examples. The new setting attribute temp_dir allows for setting where the temporary folder of an API-session should be created. By default this attribute is set to the run_dir inside the pyprodrisk installation site (maybe not the best place to create temp files, but this has been how things have worked so far, so keep this as default for backward compatibility).